activity service

+++ src/main/java/com/team5/teamscoretrackerserver/model/Activity.java
@@ -3,10 +3,12 @@ package com.team5.teamscoretrackerserver.model;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Data
 @Document
 public class Activity {
-    private Integer activityId;
+    private BigInteger activityId;
     private String activityName;
     private Integer score;
 }
